fix(desktop): remove status bar, new elements in header
